+++ entity/src/main/java/com/gestor/seguimiento/PlanSeccionDetallePK.java
@@ -19,7 +19,7 @@ public class PlanSeccionDetallePK implements Serializable {
 
     @Basic(optional = false)
     @Column(name = "codigo_establecimiento")
-    private short codigoEstablecimiento;
+    private int codigoEstablecimiento;
     @Basic(optional = false)
     @Column(name = "cod_titulo")
     private int codTitulo;
@@ -33,18 +33,18 @@ public class PlanSeccionDetallePK implements Serializable {
     public PlanSeccionDetallePK() {
     }
 
-    public PlanSeccionDetallePK(short codigoEstablecimiento, int codTitulo, int codSeccion, int codSeccionDetalle) {
+    public PlanSeccionDetallePK(int codigoEstablecimiento, int codTitulo, int codSeccion, int codSeccionDetalle) {
         this.codigoEstablecimiento = codigoEstablecimiento;
         this.codTitulo = codTitulo;
         this.codSeccion = codSeccion;
         this.codSeccionDetalle = codSeccionDetalle;
     }
 
-    public short getCodigoEstablecimiento() {
+    public int getCodigoEstablecimiento() {
         return codigoEstablecimiento;
     }
 
-    public void setCodigoEstablecimiento(short codigoEstablecimiento) {
+    public void setCodigoEstablecimiento(int codigoEstablecimiento) {
         this.codigoEstablecimiento = codigoEstablecimiento;
     }
