+++ entity/src/main/java/com/gestor/seguimiento/PlanSeccionDetalleTextoPK.java
@@ -19,7 +19,7 @@ public class PlanSeccionDetalleTextoPK implements Serializable {
 
     @Basic(optional = false)
     @Column(name = "codigo_establecimiento")
-    private short codigoEstablecimiento;
+    private int codigoEstablecimiento;
     @Basic(optional = false)
     @Column(name = "cod_titulo")
     private int codTitulo;
@@ -36,7 +36,7 @@ public class PlanSeccionDetalleTextoPK implements Serializable {
     public PlanSeccionDetalleTextoPK() {
     }
 
-    public PlanSeccionDetalleTextoPK(short codigoEstablecimiento, int codTitulo, int codSeccion, int codSeccionDetalle, int codSeccionDetalleTexto) {
+    public PlanSeccionDetalleTextoPK(int codigoEstablecimiento, int codTitulo, int codSeccion, int codSeccionDetalle, int codSeccionDetalleTexto) {
         this.codigoEstablecimiento = codigoEstablecimiento;
         this.codTitulo = codTitulo;
         this.codSeccion = codSeccion;
@@ -44,11 +44,11 @@ public class PlanSeccionDetalleTextoPK implements Serializable {
         this.codSeccionDetalleTexto = codSeccionDetalleTexto;
     }
 
-    public short getCodigoEstablecimiento() {
+    public int getCodigoEstablecimiento() {
         return codigoEstablecimiento;
     }
 
-    public void setCodigoEstablecimiento(short codigoEstablecimiento) {
+    public void setCodigoEstablecimiento(int codigoEstablecimiento) {
         this.codigoEstablecimiento = codigoEstablecimiento;
     }
